+++ src/main/java/org/terasology/gooeyDefence/movement/components/MovementComponent.java
@@ -16,14 +16,14 @@
 package org.terasology.gooeyDefence.movement.components;
 
 import org.terasology.entitySystem.Component;
-import org.terasology.math.geom.Vector3i;
+import org.terasology.math.geom.Vector3f;
 
 /**
  * Stores information on how to move the enemy.
  */
 public class MovementComponent implements Component {
     private float speed;
-    private Vector3i goal = Vector3i.zero();
+    private Vector3f goal = Vector3f.zero();
     private float reachedDistance = 0.1f;
 
     public float getSpeed() {
@@ -35,11 +35,11 @@ public class MovementComponent implements Component {
     }
 
 
-    public Vector3i getGoal() {
+    public Vector3f getGoal() {
         return goal;
     }
 
-    public void setGoal(Vector3i goal) {
+    public void setGoal(Vector3f goal) {
         this.goal = goal;
     }
